Affichage initial via le recyclerview du path des images

+++ app/src/main/java/fr/istic/hbmlh/photoloc/model/PhotoLoc.java
@@ -1,11 +1,10 @@
 package fr.istic.hbmlh.photoloc.model;
 
+import java.util.Date;
+
 import android.arch.persistence.room.ColumnInfo;
 import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.PrimaryKey;
-
-import java.util.Date;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -24,9 +23,9 @@ public class PhotoLoc {
     @ColumnInfo(name = "file_path")
     private String filePath;
 
-    private Long longitude;
+    private Double longitude;
 
-    private Long latitude;
+    private Double latitude;
 
     private Date date;
 
@@ -46,19 +45,19 @@ public class PhotoLoc {
         this.filePath = filePath;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
